agents/web-search-agent/pyproject.toml
[project]
name = "web-search-agent"
version = "0.1.0"
description = "A LangGraph-powered agent that can search the web for information."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.0a10",
    "langchain-anthropic>=1.0.0a2",
    "langchain-community>=0.3.30",
    "langchain-core>=1.0.0a6",
    "langgraph>=1.0.0a4",
    "tavily-python>=0.7.12",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.2",
]
